x/loyalty/keeper/point.go
package keeper

import (
	"context"

	"loyaltychain/x/loyalty/types"

	"cosmossdk.io/store/prefix"
	storetypes "cosmossdk.io/store/types"
	"github.com/cosmos/cosmos-sdk/runtime"
)

// RemovePoint removes a point from the store
func (k Keeper) RemovePoint(ctx sdk.Context, addr sdk.AccAddress) {
	store := ctx.KVStore(k.storeKey)
	store.Delete(types.KeyPrefix(types.PointKeyPrefix + addr.String()))
}


// GetAllPoint returns all point
func (k Keeper) GetAllPoint(ctx context.Context) (list []types.Point) {
	storeAdapter := runtime.KVStoreAdapter(k.storeService.OpenKVStore(ctx))
	store := prefix.NewStore(storeAdapter, types.KeyPrefix(types.PointKeyPrefix))
	iterator := storetypes.KVStorePrefixIterator(store, []byte{})

	defer iterator.Close()

	for ; iterator.Valid(); iterator.Next() {
		var val types.Point
		k.cdc.MustUnmarshal(iterator.Value(), &val)
		list = append(list, val)
	}

	return
}
